:memo: Update README and workplan with deployed contract details; change score type in WalletRiskScore contract

+++ contract/contracts/WalletRiskScore.sol
@@ -2,14 +2,14 @@
 pragma solidity ^0.8.28;
 
 contract WalletRiskScore {
-    mapping(address => uint8) private scores;
+    mapping(address => uint256) private scores;
 
-    function setScore(address wallet, uint8 score) external {
+    function setScore(address wallet, uint256 score) external {
         require(score <= 100, "Score must be between 0 and 100");
         scores[wallet] = score;
     }
 
-    function getScore(address wallet) external view returns (uint8) {
+    function getScore(address wallet) external view returns (uint256) {
         return scores[wallet];
     }
 }
